Jan2017 version of DMP themes
Revised set of 14 themes to be implemented for all templates in Roadmap platform. Replaces previous set of 28 used in DMPonline. Themes are used by admins to tag questions with guidance. Additional changes to themes is possible based on pilot projects to map themes to repository vocabularies and other stakeholder feedback.